Rename and refactor SQL server type mapping tests
